+++ backend/processing/llm_backends/__init__.py
@@ -1,0 +1,6 @@
+"""Per-provider LLM backend implementations.
+
+Split out of backend.processing.llm_services (a pure decomposition). The
+historical import path backend.processing.llm_services remains a facade that
+re-exports every public symbol.
+"""
