+++ backend/processing/llm_services.py
@@ -1,2693 +1,51 @@
-import json
-import logging
-import re
-from typing import Any, Dict, Generator, List, Optional, Sequence, Tuple
+"""Facade for the LLM backend implementations.
 
-from backend.utils.chat_prompt import (
-    anthropic_cached_system,
-    build_chat_context,
-    build_chat_messages,
-    build_chat_prompt,
-)
-from backend.utils.config_manager import config_manager
-from backend.utils.languages import build_output_language_prompt_section
-from backend.utils.meeting_edge import (
-    MeetingEdgeContractError,
-    MeetingEdgeRequest,
-    MeetingEdgeResult,
-    get_default_meeting_edge_prompt_template,
-)
-from backend.utils.meeting_edge import (
-    build_meeting_edge_prompt as build_meeting_edge_prompt_text,
-)
-from backend.utils.meeting_edge import (
-    build_meeting_edge_prompt_parts as build_meeting_edge_prompt_parts_text,
-)
-from backend.utils.meeting_edge import (
-    parse_meeting_edge_response as parse_meeting_edge_payload,
-)
-from backend.utils.meeting_intelligence import (
-    AutomaticMeetingIntelligenceRequest,
-    AutomaticMeetingIntelligenceResult,
-    MeetingIntelligenceContractError,
-    get_default_automatic_meeting_intelligence_prompt_template,
-)
-from backend.utils.meeting_intelligence import (
-    build_automatic_meeting_intelligence_prompt as build_automatic_meeting_intelligence_prompt_text,
-)
-from backend.utils.meeting_intelligence import (
-    finalise_automatic_meeting_intelligence_result as finalise_automatic_meeting_intelligence_payload,
-)
-from backend.utils.meeting_intelligence import (
-    parse_automatic_meeting_intelligence_response as parse_automatic_meeting_intelligence_payload,
-)
-from backend.utils.meeting_notes import (
-    MeetingEventContext,
-    append_user_notes_section,
-    build_meeting_context_prompt_section,
-    build_user_notes_prompt_section,
-)
-from backend.utils.ollama_url_policy import validate_ollama_api_url
-from backend.utils.speaker_name_suggestions import (
-    SpeakerInferenceResult,
-    parse_speaker_inference_response,
-)
-
-# Provider SDKs (openai, anthropic) are imported lazily inside the functions
-# that use them to keep heavy dependencies out of the API import path.
-
-logger = logging.getLogger(__name__)
-
-JSON_CONTRACT_ERRORS = (MeetingIntelligenceContractError, MeetingEdgeContractError)
-
-
-def summarize_llm_response_shape(response_text: str) -> Dict[str, Any]:
-    """Return non-content diagnostics for malformed structured LLM responses."""
-    stripped = response_text.strip()
-    return {
-        "chars": len(response_text),
-        "starts_with": stripped[:1],
-        "ends_with": stripped[-1:] if stripped else "",
-        "first_brace": response_text.find("{"),
-        "last_brace": response_text.rfind("}"),
-        "has_fence": "```" in response_text,
-        "has_think_tag": "<think>" in response_text or "</think>" in response_text,
-    }
-
-
-def build_eligible_speaker_labels_prompt_section(
-    eligible_labels: Optional[Sequence[str]],
-) -> str:
-    labels = [
-        str(label).strip() for label in eligible_labels or () if str(label).strip()
-    ]
-    if not labels:
-        return "All unresolved diarization labels in the transcript are eligible."
-
-    lines = ["Only return suggestions for these diarization labels:"]
-    lines.extend(f"- {label}" for label in labels)
-    return "\n".join(lines)
-
-
-class LLMBackend:
-    def infer_speaker_suggestions(
-        self,
-        transcript: str,
-        prompt_template: str = None,
-        timeout: int = 60,
-        user_notes: Optional[str] = None,
-        meeting_context: Optional[MeetingEventContext] = None,
-        eligible_labels: Optional[Sequence[str]] = None,
-    ) -> SpeakerInferenceResult:
-        """
-        Infer evidence-backed name suggestions for unresolved diarization labels.
-        """
-        raise NotImplementedError
-
-    def infer_speakers(
-        self,
-        transcript: str,
-        prompt_template: str = None,
-        timeout: int = 60,
-        user_notes: Optional[str] = None,
-        meeting_context: Optional[MeetingEventContext] = None,
-        eligible_labels: Optional[Sequence[str]] = None,
-    ) -> Dict[str, str]:
-        """
-        Infer the most likely real names or roles for each speaker label in the transcript.
-        Returns a mapping from diarization label to inferred name/role.
-        """
-        result = self.infer_speaker_suggestions(
-            transcript,
-            prompt_template,
-            timeout,
-            user_notes=user_notes,
-            meeting_context=meeting_context,
-            eligible_labels=eligible_labels,
-        )
-        return result.mapping
-
-    def list_models(self) -> List[str]:
-        """
-        List available models for the provider.
-        """
-        raise NotImplementedError
-
-    def generate_meeting_notes(
-        self,
-        transcript: str,
-        speaker_mapping: Dict[str, str],
-        prompt_template: str = None,
-        timeout: int = 60,
-        user_notes: Optional[str] = None,
-        meeting_context: Optional[MeetingEventContext] = None,
-        output_language_instruction: Optional[str] = None,
-    ) -> str:
-        """
-        Generate meeting notes using the provided speaker mapping to replace generic labels.
-        Returns the meeting notes as a string.
-        """
-        raise NotImplementedError
-
-    def generate_meeting_intelligence(
-        self,
-        request: AutomaticMeetingIntelligenceRequest,
-        prompt_template: str = None,
-        timeout: int = 60,
-    ) -> AutomaticMeetingIntelligenceResult:
-        """
-        Generate speaker suggestions for unresolved labels, a meeting title, and
-        meeting notes from a single LLM call.
-        """
-        raise NotImplementedError
-
-    def generate_meeting_edge(
-        self,
-        request: MeetingEdgeRequest,
-        prompt_template: str = None,
-        timeout: int = 60,
-    ) -> MeetingEdgeResult:
-        """
-        Generate live Meeting Edge guidance from the current meeting context.
-        """
-        raise NotImplementedError
-
-    def infer_speakers_and_generate_notes(
-        self,
-        transcript: str,
-        prompt_template: str = None,
-        timeout: int = 60,
-        user_notes: Optional[str] = None,
-    ) -> Tuple[Dict[str, str], str]:
-        """
-        Backward-compatible method: infers speakers and generates notes in sequence using the new methods.
-        """
-        mapping = self.infer_speakers(
-            transcript,
-            prompt_template,
-            timeout,
-            user_notes=user_notes,
-        )
-        notes = self.generate_meeting_notes(
-            transcript, mapping, prompt_template, timeout, user_notes=user_notes
-        )
-        return mapping, notes
-
-    def ask_question_about_meeting(
-        self,
-        user_question: str,
-        meeting_notes: str,
-        diarized_transcript: str,
-        conversation_history: list = None,
-        timeout: int = 60,
-        recording_id: str = None,
-    ):
-        """
-        Ask a question about the meeting.
-        """
-        raise NotImplementedError
-
-    def ask_question_streaming(
-        self,
-        user_question: str,
-        meeting_notes: str,
-        diarized_transcript: str,
-        conversation_history: list = None,
-        timeout: int = 60,
-        recording_id: str = None,
-    ) -> Generator[str, None, None]:
-        """
-        Ask a question about the meeting and yield response chunks.
-        """
-        raise NotImplementedError
-
-    def infer_meeting_title(
-        self,
-        transcript: str,
-        prompt_template: str = None,
-        timeout: int = 60,
-        output_language_instruction: Optional[str] = None,
-    ) -> str:
-        """
-        Infer a concise, descriptive meeting title from the provided transcript.
-        Sub-classes must implement.
-        """
-        raise NotImplementedError
-
-    def validate_api_key(self) -> bool:
-        """
-        Validate the API key by making a lightweight API call.
-        Returns True if valid, raises an exception or returns False if invalid.
-        """
-        raise NotImplementedError
-
-    def _build_chat_prompt(
-        self, user_question: str, meeting_notes: str, diarized_transcript: str
-    ) -> str:
-        return build_chat_prompt(user_question, meeting_notes, diarized_transcript)
-
-    @staticmethod
-    def get_default_speaker_prompt_template():
-        return """
-You are an expert meeting assistant. Analyze the diarized meeting transcript below, where speakers are labeled generically (e.g., 'Speaker 1', 'SPEAKER_00').\n\nFirst, infer the most likely real names or roles for each speaker, based on context, introductions, or references in the transcript. Use the user-authored notes as supplemental context when they mention attendee names, roles, or who said what. If a real name is not clear, suggest a likely role (e.g., 'Project Manager', 'Client', 'Engineer') or keep the generic label. Be conservative: only use a real name or role if it is clearly stated or strongly implied. If the user notes conflict with the transcript, prefer the transcript.\n\nOutput a Markdown table mapping each diarization label to the inferred name or role. Only output the table and nothing else.\n\n# User Notes Context\n{user_notes_section}\n\n# Meeting Context\n{meeting_context_section}\n\nBelow is the diarized transcript:\n\n{transcript}\n"""
-
-    @staticmethod
-    def get_default_speaker_suggestion_prompt_template():
-        return """
-You are an expert meeting assistant. Analyze the diarized meeting transcript below and return only evidence-backed speaker name suggestions for unresolved diarization labels.
-
-# Critical Rules
-- Return valid JSON only. Do not include Markdown, commentary, or code fences unless the client asks for them.
-- Only include suggestions for the eligible diarization labels listed below.
-- Omit a label entirely if the transcript evidence is weak or ambiguous.
-- Each suggestion must include at least one direct evidence span quoted from the transcript.
-- Prefer names already present in the transcript or linked meeting attendees. Do not invent names.
-- Be conservative. If you are unsure, omit the suggestion instead of guessing.
-- `confidence` must be between 0.0 and 1.0.
-
-# Eligible Diarization Labels
-{eligible_labels_section}
-
-# User Notes Context
-{user_notes_section}
-
-# Meeting Context
-{meeting_context_section}
-
-# Required JSON Schema
-{{
-    "suggestions": [
-        {{
-            "diarization_label": "SPEAKER_00",
-            "suggested_name": "Alex Johnson",
-            "confidence": 0.93,
-            "rationale": "The speaker says 'I'm Alex' and the attendee list contains Alex Johnson.",
-            "signals": ["self_introduction", "meeting_attendee_exact"],
-            "evidence_spans": [
-                {{
-                    "quote": "Hi everyone, I'm Alex from product.",
-                    "reason": "self_introduction",
-                    "start_seconds": 0.0,
-                    "end_seconds": 3.2
-                }}
-            ]
-        }}
-    ]
-}}
-
-# Transcript
-{transcript}
+The per-provider backends, shared base class, and factory helpers now live in
+``backend.processing.llm_backends.*``. This module is retained as a thin
+compatibility facade so the historical import path
+``backend.processing.llm_services`` keeps working unchanged for every call site.
+This is a pure decomposition: no behaviour, signatures, or prompt text changed.
 """
 
-    @staticmethod
-    def get_default_notes_prompt_template():
-        return """You are an expert meeting intelligence assistant. Your task is to generate comprehensive, high-quality meeting notes from the provided transcript. Use the speaker mapping to refer to participants by their inferred names/roles instead of generic labels.
-
-# CRITICAL FORMATTING REQUIREMENTS
-You MUST follow these formatting rules EXACTLY. Do not deviate:
-1. Use ONLY the section structure specified below, with headings translated into the requested output language
-2. Use Markdown formatting throughout
-3. Be thorough and detailed - notes may be as lengthy as required to capture all important content
-4. Do NOT add any introductory text, concluding remarks, or sections not specified below
-5. Start your response with one top-level Markdown heading (`# ...`) in the requested output language - nothing before it
-6. Incorporate relevant user-authored notes into the appropriate sections when they add useful context or action items
-7. Do NOT add a separate appendix or label LLM-generated content as such; the application will surface the user-authored notes separately
-
-# Speaker Mapping
-{mapping_table}
-
-# User Notes Context
-{user_notes_section}
-
-# Meeting Context
-{meeting_context_section}
-
-{output_language_section}
-
-# OUTPUT FORMAT - Follow this EXACT structure:
-
-# [Localized meeting-notes heading]
-
-## [Localized Topics Discussed heading]
-List each major topic or theme discussed in the meeting as a bullet point. Be specific and descriptive.
-- Topic 1: Brief description
-- Topic 2: Brief description
-(continue for all topics)
-
-## [Localized Summary heading]
-Provide a comprehensive summary of the meeting covering:
-- The main purpose and context of the meeting
-- Key points raised by participants
-- Important information shared
-- Overall conclusions or outcomes reached
-
-## [Localized Detailed Notes heading]
-
-### [Topic Name 1]
-Provide detailed notes on this topic including:
-- **Key Points**: Main arguments, information, or ideas presented
-- **Discussion**: What was debated or discussed, including different perspectives
-- **Decisions**: Any decisions made regarding this topic (if applicable)
-- **Rationale**: The reasoning behind decisions or recommendations (if discussed)
-- **Open Questions**: Any unresolved questions or points requiring follow-up
-
-### [Topic Name 2]
-(Follow the same structure for each major topic)
-
-(Continue for all topics discussed)
-
-## [Localized Action Items / Tasks heading]
-List all tasks, action items, or follow-ups mentioned, formatted as:
-- [ ] Task description - Assigned to: [Person] - Due: [Date if mentioned, otherwise "TBD"]
-(If no tasks were discussed, write: "No specific action items were identified in this meeting.")
-
-## [Localized Miscellaneous heading]
-Capture any additional important information that doesn't fit the above categories:
-- Side discussions or tangential points of interest
-- Announcements or FYIs mentioned
-- References to external documents, resources, or prior meetings
-- Any other noteworthy content
-(If nothing applicable, write: "No additional items.")
-
----
-
-# Transcript to Analyze:
-
-{transcript}
-
----
-
-Now generate the meeting notes following the exact structure specified above, replacing bracketed heading descriptions with headings in the requested output language. Be comprehensive and capture all important details."""
-
-    @staticmethod
-    def get_default_title_prompt_template():
-        return (
-            "You are an expert meeting assistant. Given the full meeting transcript below, "
-            "provide a concise, descriptive title that summarises the main topic or purpose of the meeting. "
-            "Limit the title to at most 12 words. Output ONLY the title with no additional commentary, punctuation, or formatting.\n\n"
-            "{output_language_section}\n\n"
-            "# Transcript\n\n{transcript}\n"
-        )
-
-    @staticmethod
-    def get_speaker_prompt_template() -> str:
-        return LLMBackend.get_default_speaker_prompt_template()
-
-    @staticmethod
-    def get_notes_prompt_template() -> str:
-        return LLMBackend.get_default_notes_prompt_template()
-
-    @staticmethod
-    def get_title_prompt_template() -> str:
-        return LLMBackend.get_default_title_prompt_template()
-
-    @staticmethod
-    def get_speaker_suggestion_prompt_template() -> str:
-        return LLMBackend.get_default_speaker_suggestion_prompt_template()
-
-    @staticmethod
-    def get_automatic_meeting_intelligence_prompt_template() -> str:
-        return get_default_automatic_meeting_intelligence_prompt_template()
-
-    @staticmethod
-    def get_meeting_edge_prompt_template() -> str:
-        return get_default_meeting_edge_prompt_template()
-
-    @staticmethod
-    def parse_mapping_table(response_text: str) -> Dict[str, str]:
-        lines = [line.strip() for line in response_text.splitlines() if line.strip()]
-        mapping = {}
-        in_table = False
-        for line in lines:
-            if line.startswith("|") and "|" in line[1:]:
-                in_table = True
-                parts = [p.strip() for p in line.strip("|").split("|")]
-                if (
-                    len(parts) == 2
-                    and parts[0] != "Diarization Label"
-                    and not parts[0].startswith("-")
-                ):
-                    mapping[parts[0]] = parts[1]
-            elif in_table and (not line.startswith("|")):
-                break
-        return mapping
-
-    @staticmethod
-    def parse_notes(response_text: str) -> str:
-        # Assume notes start after the mapping table (after a blank line or after '# Meeting Notes')
-        lines = [line for line in response_text.splitlines()]
-        notes_lines = []
-        in_notes = False
-        for line in lines:
-            if line.strip().startswith("# Meeting Notes"):
-                in_notes = True
-            if in_notes:
-                notes_lines.append(line)
-
-        result = "\n".join(notes_lines).strip()
-        if not result:
-            # Fallback: if no header found, return everything.
-            # This handles cases where the prompt is modified or the LLM disobeys.
-            return response_text.strip()
-        return result
-
-    @staticmethod
-    def parse_title(response_text: str) -> str:
-        # Take first non-empty line, strip spurious characters/quotes/markdown
-        for line in response_text.splitlines():
-            cleaned = line.strip().lstrip("#").strip()
-            cleaned = re.sub(
-                r"^\W+|\W+$", "", cleaned
-            )  # Remove leading/trailing non-word chars/quotes
-            if cleaned:
-                cleaned = re.sub(r"\s+", " ", cleaned)
-                return cleaned
-        return response_text.strip()
-
-    @staticmethod
-    def mapping_to_markdown_table(mapping: Dict[str, str]) -> str:
-        if not mapping:
-            return ""
-        header = "| Diarization Label | Inferred Name/Role |\n|---|---|"
-        rows = [f"| {k} | {v} |" for k, v in mapping.items()]
-        return "\n".join([header] + rows)
-
-    @staticmethod
-    def build_notes_prompt(
-        prompt_template: str,
-        transcript: str,
-        speaker_mapping: Dict[str, str],
-        user_notes: Optional[str] = None,
-        meeting_context: Optional[MeetingEventContext] = None,
-        output_language_instruction: Optional[str] = None,
-    ) -> str:
-        return prompt_template.format(
-            transcript=transcript,
-            mapping_table=LLMBackend.mapping_to_markdown_table(speaker_mapping),
-            user_notes_section=build_user_notes_prompt_section(user_notes),
-            meeting_context_section=build_meeting_context_prompt_section(
-                meeting_context
-            ),
-            output_language_section=build_output_language_prompt_section(
-                output_language_instruction
-            ),
-        )
-
-    @staticmethod
-    def build_speaker_prompt(
-        prompt_template: str,
-        transcript: str,
-        user_notes: Optional[str] = None,
-        meeting_context: Optional[MeetingEventContext] = None,
-    ) -> str:
-        return prompt_template.format(
-            transcript=transcript,
-            user_notes_section=build_user_notes_prompt_section(user_notes),
-            meeting_context_section=build_meeting_context_prompt_section(
-                meeting_context
-            ),
-        )
-
-    @staticmethod
-    def build_speaker_suggestion_prompt(
-        prompt_template: str,
-        transcript: str,
-        eligible_labels: Optional[Sequence[str]] = None,
-        user_notes: Optional[str] = None,
-        meeting_context: Optional[MeetingEventContext] = None,
-    ) -> str:
-        return prompt_template.format(
-            transcript=transcript,
-            eligible_labels_section=build_eligible_speaker_labels_prompt_section(
-                eligible_labels
-            ),
-            user_notes_section=build_user_notes_prompt_section(user_notes),
-            meeting_context_section=build_meeting_context_prompt_section(
-                meeting_context
-            ),
-        )
-
-    @staticmethod
-    def build_automatic_meeting_intelligence_prompt(
-        request: AutomaticMeetingIntelligenceRequest,
-        prompt_template: str = None,
-    ) -> str:
-        return build_automatic_meeting_intelligence_prompt_text(
-            request,
-            prompt_template,
-        )
-
-    @staticmethod
-    def build_meeting_edge_prompt(
-        request: MeetingEdgeRequest,
-        prompt_template: str = None,
-    ) -> str:
-        return build_meeting_edge_prompt_text(request, prompt_template)
-
-    @staticmethod
-    def build_meeting_edge_prompt_parts(
-        request: MeetingEdgeRequest,
-        prompt_template: str = None,
-    ) -> Tuple[str, str]:
-        return build_meeting_edge_prompt_parts_text(request, prompt_template)
-
-    @staticmethod
-    def finalise_meeting_notes(notes: str, user_notes: Optional[str] = None) -> str:
-        return append_user_notes_section(notes, user_notes)
-
-    @staticmethod
-    def parse_automatic_meeting_intelligence_result(
-        response_text: str,
-        request: AutomaticMeetingIntelligenceRequest,
-    ) -> AutomaticMeetingIntelligenceResult:
-        result = parse_automatic_meeting_intelligence_payload(
-            response_text,
-            request=request,
-        )
-        return finalise_automatic_meeting_intelligence_payload(
-            result,
-            request.user_notes,
-        )
-
-    @staticmethod
-    def parse_speaker_inference_result(
-        response_text: str,
-        allowed_labels: Optional[Sequence[str]] = None,
-    ) -> SpeakerInferenceResult:
-        return parse_speaker_inference_response(
-            response_text,
-            allowed_labels=allowed_labels,
-        )
-
-    @staticmethod
-    def parse_meeting_edge_result(
-        response_text: str,
-        request: MeetingEdgeRequest,
-    ) -> MeetingEdgeResult:
-        return parse_meeting_edge_payload(response_text, request=request)
-
-    @staticmethod
-    def build_json_repair_prompt(
-        *,
-        original_prompt: str,
-        invalid_response: str,
-        validation_error: Exception,
-    ) -> str:
-        return f"""Your previous response did not satisfy Nojoin's strict JSON contract.
-
-Validation error:
-{validation_error}
-
-Return a corrected response for the original task as one valid JSON object only.
-Do not include Markdown code fences, commentary, or prose before or after the JSON object.
-Preserve the same schema and do not invent facts not supported by the original task.
-
-# Original Task
-{original_prompt}
-
-# Previous Invalid Response
-{invalid_response}
-"""
-
-    @staticmethod
-    def get_mapped_transcript_for_llm(recording_id: int) -> str:
-        """
-        Fetches the diarized transcript and speaker mapping for a recording, and returns the mapped transcript as plaintext.
-        """
-        from sqlmodel import select
-
-        from backend.core.db import get_sync_session
-        from backend.models.recording import Recording
-        from backend.models.speaker import RecordingSpeaker
-        from backend.models.transcript import Transcript
-
-        with get_sync_session() as session:
-            rec = session.get(Recording, recording_id)
-            if not rec:
-                return "Recording not found."
-
-            # Get Transcript
-            transcript_obj = session.exec(
-                select(Transcript).where(Transcript.recording_id == recording_id)
-            ).first()
-            if not transcript_obj or not transcript_obj.segments:
-                return "Diarized transcript not found."
-
-            # Get Speakers
-            speakers = session.exec(
-                select(RecordingSpeaker).where(
-                    RecordingSpeaker.recording_id == recording_id
-                )
-            ).all()
-            label_to_name = {s.diarization_label: s.name for s in speakers}
-
-            # Render
-            lines = []
-            for seg in transcript_obj.segments:
-                speaker_label = seg.get("speaker", "Unknown")
-                speaker_name = label_to_name.get(speaker_label, speaker_label)
-                text = seg.get("text", "")
-                start = seg.get("start", 0)
-                minutes = int(start // 60)
-                seconds = int(start % 60)
-                timestamp = f"[{minutes:02d}:{seconds:02d}]"
-                lines.append(f"{timestamp} {speaker_name}: {text}")
-
-            return "\n".join(lines)
-
-    def _update_notes_in_db(self, recording_id: int, new_notes: str):
-        """
-        Updates the meeting notes in the database.
-        """
-        from sqlmodel import select
-
-        from backend.core.db import get_sync_session
-        from backend.models.transcript import Transcript
-
-        with get_sync_session() as session:
-            transcript_obj = session.exec(
-                select(Transcript).where(Transcript.recording_id == recording_id)
-            ).first()
-            if transcript_obj:
-                transcript_obj.notes = new_notes
-                session.add(transcript_obj)
-                session.commit()
-                logger.info(f"Updated notes for recording {recording_id}")
-            else:
-                logger.error(
-                    f"Could not find transcript for recording {recording_id} to update notes"
-                )
-
-
-class GeminiLLMBackend(LLMBackend):
-    def __init__(self, api_key=None, model=None):
-        # Lazy import to avoid errors when google-genai isn't installed
-        try:
-            from google import genai
-        except ImportError:
-            raise ImportError(
-                "The 'google-genai' package is required for Gemini support. "
-                "Please install it with: pip install google-genai"
-            )
-
-        if api_key is None:
-            api_key = config_manager.get("gemini_api_key")
-        if not api_key:
-            raise ValueError(
-                "Google Gemini API key is not set. Please provide it in settings."
-            )
-        self.api_key = api_key
-        self.model = model or _get_default_model_for_provider("gemini")
-        self.genai = genai  # Store reference for later use
-        self.client = genai.Client(api_key=self.api_key)
-
-    def _extract_text_from_response(self, response):
-        """
-        Extract text from the response, handling potential non-text parts (like thoughts)
-        to avoid warnings and ensure text extraction.
-        """
-        if hasattr(response, "candidates") and response.candidates:
-            candidate = response.candidates[0]
-            if (
-                hasattr(candidate, "content")
-                and candidate.content
-                and hasattr(candidate.content, "parts")
-            ):
-                text_parts = []
-                for part in candidate.content.parts:
-                    if hasattr(part, "text") and part.text:
-                        text_parts.append(part.text)
-                if text_parts:
-                    return "".join(text_parts)
-
-        # Fallback to .text if available (which might log the warning but works)
-        if hasattr(response, "text"):
-            return response.text
-        return ""
-
-    def list_models(self) -> List[str]:
-        """
-        List available Gemini models.
-        """
-        try:
-            models = self.client.models.list()
-            # Extract model IDs (e.g. 'models/gemini-pro') from the API response.
-            model_list = []
-            for m in models:
-                # Check attributes
-                name = getattr(m, "name", None)
-                if name:
-                    # Strip 'models/' prefix if present, as the client usually handles it or expects it.
-                    # But for display we might want the full name or short name.
-                    # The user example showed 'gemini-flash-latest'.
-                    if name.startswith("models/"):
-                        name = name[7:]
-
-                    # Filter for gemini models
-                    if "gemini" in name.lower():
-                        model_list.append(name)
-
-            return sorted(model_list)
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Gemini API error (list models): {e}")
-            return []
-
-    def infer_speaker_suggestions(
-        self,
-        transcript: str,
-        prompt_template: str = None,
-        timeout: int = 60,
-        user_notes: Optional[str] = None,
-        meeting_context: Optional[MeetingEventContext] = None,
-        eligible_labels: Optional[Sequence[str]] = None,
-    ) -> SpeakerInferenceResult:
-        """
-        Run speaker inference on the transcript and return structured suggestions.
-        Can be called independently of meeting notes generation.
-        """
-        if prompt_template is None:
-            prompt_template = self.get_speaker_suggestion_prompt_template()
-        prompt = self.build_speaker_suggestion_prompt(
-            prompt_template,
-            transcript,
-            eligible_labels,
-            user_notes,
-            meeting_context,
-        )
-        if not self.model:
-            raise ValueError(
-                "No Gemini model configured. Please select a model in Settings."
-            )
-        try:
-            response = self.client.models.generate_content(
-                model=self.model,
-                contents=prompt,
-            )
-            text = self._extract_text_from_response(response)
-            return self.parse_speaker_inference_result(text, eligible_labels)
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Gemini API error (speaker suggestions): {e}")
-            raise RuntimeError(f"Gemini API error (speaker suggestions): {e}")
-
-    def infer_speakers(
-        self,
-        transcript: str,
-        prompt_template: str = None,
-        timeout: int = 60,
-        user_notes: Optional[str] = None,
-        meeting_context: Optional[MeetingEventContext] = None,
-        eligible_labels: Optional[Sequence[str]] = None,
-    ) -> Dict[str, str]:
-        return self.infer_speaker_suggestions(
-            transcript,
-            prompt_template,
-            timeout,
-            user_notes=user_notes,
-            meeting_context=meeting_context,
-            eligible_labels=eligible_labels,
-        ).mapping
-
-    def generate_meeting_notes(
-        self,
-        transcript: str,
-        speaker_mapping: Dict[str, str],
-        prompt_template: str = None,
-        timeout: int = 60,
-        user_notes: Optional[str] = None,
-        meeting_context: Optional[MeetingEventContext] = None,
-        output_language_instruction: Optional[str] = None,
-    ) -> str:
-        """
-        Generate meeting notes using the provided speaker mapping. Should be called after user relabeling.
-        """
-        if prompt_template is None:
-            prompt_template = self.get_notes_prompt_template()
-        prompt = self.build_notes_prompt(
-            prompt_template,
-            transcript,
-            speaker_mapping,
-            user_notes,
-            meeting_context,
-            output_language_instruction,
-        )
-        if not self.model:
-            raise ValueError(
-                "No Gemini model configured. Please select a model in Settings."
-            )
-        try:
-            response = self.client.models.generate_content(
-                model=self.model,
-                contents=prompt,
-            )
-            text = self._extract_text_from_response(response)
-            notes = self.finalise_meeting_notes(self.parse_notes(text), user_notes)
-            return notes
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Gemini API error (meeting notes): {e}")
-            raise RuntimeError(f"Gemini API error (meeting notes): {e}")
-
-    def generate_meeting_intelligence(
-        self,
-        request: AutomaticMeetingIntelligenceRequest,
-        prompt_template: str = None,
-        timeout: int = 60,
-    ) -> AutomaticMeetingIntelligenceResult:
-        prompt = self.build_automatic_meeting_intelligence_prompt(
-            request,
-            prompt_template,
-        )
-        if not self.model:
-            raise ValueError(
-                "No Gemini model configured. Please select a model in Settings."
-            )
-        try:
-            response = self.client.models.generate_content(
-                model=self.model,
-                contents=prompt,
-            )
-            text = self._extract_text_from_response(response)
-            return self.parse_automatic_meeting_intelligence_result(text, request)
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Gemini API error (meeting intelligence): {e}")
-            raise RuntimeError(f"Gemini API error (meeting intelligence): {e}")
-
-    def generate_meeting_edge(
-        self,
-        request: MeetingEdgeRequest,
-        prompt_template: str = None,
-        timeout: int = 60,
-    ) -> MeetingEdgeResult:
-        prompt = self.build_meeting_edge_prompt(request, prompt_template)
-        if not self.model:
-            raise ValueError(
-                "No Gemini model configured. Please select a model in Settings."
-            )
-        try:
-            response = self.client.models.generate_content(
-                model=self.model,
-                contents=prompt,
-                config=self.genai.types.GenerateContentConfig(
-                    response_mime_type="application/json",
-                    temperature=0.2,
-                    http_options=self.genai.types.HttpOptions(
-                        timeout=int(timeout * 1000),
-                    ),
-                ),
-            )
-            text = self._extract_text_from_response(response)
-            return self.parse_meeting_edge_result(text, request)
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Gemini API error (Meeting Edge): {e}")
-            raise RuntimeError(f"Gemini API error (Meeting Edge): {e}")
-
-    # infer_speakers_and_generate_notes is inherited and calls the above two methods
-
-    def ask_question_about_meeting(
-        self,
-        user_question: str,
-        meeting_notes: str,
-        diarized_transcript: str,
-        conversation_history: list = None,
-        timeout: int = 60,
-        recording_id: str = None,
-    ):
-        # If recording_id is provided, use mapped transcript
-        if recording_id is not None:
-            diarized_transcript = self.get_mapped_transcript_for_llm(recording_id)
-
-        prompt = self._build_chat_prompt(
-            user_question, meeting_notes, diarized_transcript
-        )
-
-        contents = []
-        if conversation_history:
-            contents.extend(conversation_history)
-        contents.append({"role": "user", "parts": [{"text": prompt}]})
-        if not self.model:
-            raise ValueError(
-                "No Gemini model configured. Please select a model in Settings."
-            )
-        try:
-            response = self.client.models.generate_content(
-                model=self.model,
-                contents=contents,
-            )
-            return self._extract_text_from_response(response)
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Gemini API error (chat): {e}")
-            raise RuntimeError(f"Gemini API error (chat): {e}")
-
-    def ask_question_streaming(
-        self,
-        user_question: str,
-        meeting_notes: str,
-        diarized_transcript: str,
-        conversation_history: list = None,
-        timeout: int = 60,
-        recording_id: str = None,
-    ) -> Generator[Any, None, None]:
-        if recording_id is not None:
-            diarized_transcript = self.get_mapped_transcript_for_llm(recording_id)
-
-        prompt = self._build_chat_prompt(
-            user_question, meeting_notes, diarized_transcript
-        )
-
-        contents = []
-        if conversation_history:
-            contents.extend(conversation_history)
-        contents.append({"role": "user", "parts": [{"text": prompt}]})
-
-        if not self.model:
-            raise ValueError(
-                "No Gemini model configured. Please select a model in Settings."
-            )
-
-        # Define Tool
-        def update_meeting_notes(content: str):
-            """Overwrites the current meeting notes with new content. Use this whenever the user asks to modify, edit, add to, or delete parts of the meeting notes. The input should be the fully updated Markdown content of the notes."""
-            pass
-
-        tools = [update_meeting_notes]
-
-        try:
-            # Use streaming API
-            # Automatic function calling is disabled to allow manual handling and event streaming.
-            response_stream = self.client.models.generate_content_stream(
-                model=self.model,
-                contents=contents,
-                config=self.genai.types.GenerateContentConfig(
-                    tools=tools,
-                    automatic_function_calling=self.genai.types.AutomaticFunctionCallingConfig(
-                        disable=True
-                    ),
-                ),
-            )
-            for chunk in response_stream:
-                # Check for function calls
-                if chunk.function_calls:
-                    for fc in chunk.function_calls:
-                        if fc.name == "update_meeting_notes":
-                            new_notes = fc.args.get("content")
-                            if new_notes and recording_id:
-                                self._update_notes_in_db(recording_id, new_notes)
-                                yield {"type": "notes_update"}
-                                # Send a confirmation message back to the user since we are not doing a full round-trip
-                                yield "I have updated the meeting notes."
-
-                # Safely extract text to avoid warnings about non-text parts
-                try:
-                    # Check if there is actual text content before accessing .text
-                    has_text = False
-                    if hasattr(chunk, "candidates") and chunk.candidates:
-                        candidate = chunk.candidates[0]
-                        if (
-                            hasattr(candidate, "content")
-                            and candidate.content
-                            and hasattr(candidate.content, "parts")
-                        ):
-                            for part in candidate.content.parts:
-                                if hasattr(part, "text") and part.text:
-                                    has_text = True
-                                    break
-
-                    if has_text and chunk.text:
-                        yield chunk.text
-                except Exception:  # noqa: BLE001
-                    pass
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Gemini API error (streaming chat): {e}")
-            # If it's just a property access error because of no text, ignore it
-            if "Candidate was blocked" in str(e) or "has no parts" in str(e):
-                pass
-            else:
-                raise RuntimeError(f"Gemini API error (streaming chat): {e}")
-
-    def infer_meeting_title(
-        self,
-        transcript: str,
-        prompt_template: str = None,
-        timeout: int = 60,
-        output_language_instruction: Optional[str] = None,
-    ) -> str:
-        """
-        Infer a concise, descriptive meeting title from the provided transcript.
-        Sub-classes must implement.
-        """
-        if prompt_template is None:
-            prompt_template = self.get_title_prompt_template()
-        prompt = prompt_template.format(
-            transcript=transcript,
-            output_language_section=build_output_language_prompt_section(
-                output_language_instruction
-            ),
-        )
-        if not self.model:
-            raise ValueError(
-                "No Gemini model configured. Please select a model in Settings."
-            )
-        try:
-            response = self.client.models.generate_content(
-                model=self.model,
-                contents=prompt,
-            )
-            text = self._extract_text_from_response(response)
-            title = self.parse_title(text)
-            return title
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Gemini API error (meeting title): {e}")
-            raise RuntimeError(f"Gemini API error (meeting title): {e}")
-
-    def validate_api_key(self) -> bool:
-        """
-        Validate the API key by making a lightweight API call.
-        Returns True if valid, raises an exception or returns False if invalid.
-        """
-        try:
-            # Simple call to list models to verify key
-            self.client.models.list()
-            return True
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Gemini API validation failed: {e}")
-            raise ValueError(f"Gemini API validation failed: {e}")
-
-
-class OpenAILLMBackend(LLMBackend):
-    def __init__(self, api_key=None, model=None):
-        import openai
-
-        if api_key is None:
-            api_key = config_manager.get("openai_api_key")
-        if not api_key:
-            raise ValueError(
-                "OpenAI API key is not set. Please provide it in settings."
-            )
-        self.api_key = api_key
-        self.model = model or _get_default_model_for_provider("openai")
-        self.client = openai.OpenAI(api_key=self.api_key)
-
-    def list_models(self) -> List[str]:
-        """
-        List available OpenAI models.
-        """
-        try:
-            models = self.client.models.list()
-            # Filter for gpt models to avoid clutter (dall-e, whisper, etc)
-            # Also exclude models not supported by chat endpoint if possible, but it's hard to know for sure.
-            model_list = []
-            for m in models:
-                if "gpt" in m.id and "audio" not in m.id and "realtime" not in m.id:
-                    model_list.append(m.id)
-                elif "o1" in m.id:  # Add support for reasoning models
-                    model_list.append(m.id)
-            return sorted(model_list)
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"OpenAI API error (list models): {e}")
-            return []
-
-    def infer_speaker_suggestions(
-        self,
-        transcript: str,
-        prompt_template: str = None,
-        timeout: int = 60,
-        user_notes: Optional[str] = None,
-        meeting_context: Optional[MeetingEventContext] = None,
-        eligible_labels: Optional[Sequence[str]] = None,
-    ) -> SpeakerInferenceResult:
-        """
-        Run speaker inference on the transcript and return structured suggestions.
-        Can be called independently of meeting notes generation.
-        """
-        if prompt_template is None:
-            prompt_template = self.get_speaker_suggestion_prompt_template()
-        prompt = self.build_speaker_suggestion_prompt(
-            prompt_template,
-            transcript,
-            eligible_labels,
-            user_notes,
-            meeting_context,
-        )
-        if not self.model:
-            raise ValueError(
-                "No OpenAI model configured. Please select a model in Settings."
-            )
-        try:
-            stream = self.client.chat.completions.create(
-                model=self.model,
-                messages=[{"role": "user", "content": prompt}],
-                temperature=0.2,
-                timeout=timeout,
-                stream=True,
-            )
-            text_chunks = []
-            for chunk in stream:
-                if chunk.choices[0].delta.content:
-                    text_chunks.append(chunk.choices[0].delta.content)
-            text = "".join(text_chunks)
-            return self.parse_speaker_inference_result(text, eligible_labels)
-        except Exception as e:  # noqa: BLE001
-            if "not a chat model" in str(e) or "404" in str(e):
-                logger.error(
-                    f"OpenAI API error (speaker suggestions): Invalid model {self.model}. {e}"
-                )
-                raise ValueError(
-                    f"The model '{self.model}' appears to be invalid or is not a chat model. Please check the model name in Settings."
-                )
-            logger.error(f"OpenAI API error (speaker suggestions): {e}")
-            raise RuntimeError(f"OpenAI API error (speaker suggestions): {e}")
-
-    def infer_speakers(
-        self,
-        transcript: str,
-        prompt_template: str = None,
-        timeout: int = 60,
-        user_notes: Optional[str] = None,
-        meeting_context: Optional[MeetingEventContext] = None,
-        eligible_labels: Optional[Sequence[str]] = None,
-    ) -> Dict[str, str]:
-        return self.infer_speaker_suggestions(
-            transcript,
-            prompt_template,
-            timeout,
-            user_notes=user_notes,
-            meeting_context=meeting_context,
-            eligible_labels=eligible_labels,
-        ).mapping
-
-    def generate_meeting_notes(
-        self,
-        transcript: str,
-        speaker_mapping: Dict[str, str],
-        prompt_template: str = None,
-        timeout: int = 60,
-        user_notes: Optional[str] = None,
-        meeting_context: Optional[MeetingEventContext] = None,
-        output_language_instruction: Optional[str] = None,
-    ) -> str:
-        """
-        Generate meeting notes using the provided speaker mapping. Should be called after user relabeling.
-        """
-        if prompt_template is None:
-            prompt_template = self.get_notes_prompt_template()
-        prompt = self.build_notes_prompt(
-            prompt_template,
-            transcript,
-            speaker_mapping,
-            user_notes,
-            meeting_context,
-            output_language_instruction,
-        )
-        if not self.model:
-            raise ValueError(
-                "No OpenAI model configured. Please select a model in Settings."
-            )
-        try:
-            stream = self.client.chat.completions.create(
-                model=self.model,
-                messages=[{"role": "user", "content": prompt}],
-                temperature=0.2,
-                timeout=timeout,
-                stream=True,
-            )
-            text_chunks = []
-            for chunk in stream:
-                if chunk.choices[0].delta.content:
-                    text_chunks.append(chunk.choices[0].delta.content)
-            text = "".join(text_chunks)
-            notes = self.finalise_meeting_notes(self.parse_notes(text), user_notes)
-            return notes
-        except Exception as e:  # noqa: BLE001
-            if "not a chat model" in str(e) or "404" in str(e):
-                logger.error(
-                    f"OpenAI API error (meeting notes): Invalid model {self.model}. {e}"
-                )
-                raise ValueError(
-                    f"The model '{self.model}' appears to be invalid or is not a chat model. Please check the model name in Settings."
-                )
-            logger.error(f"OpenAI API error (meeting notes): {e}")
-            raise RuntimeError(f"OpenAI API error (meeting notes): {e}")
-
-    def generate_meeting_intelligence(
-        self,
-        request: AutomaticMeetingIntelligenceRequest,
-        prompt_template: str = None,
-        timeout: int = 60,
-    ) -> AutomaticMeetingIntelligenceResult:
-        prompt = self.build_automatic_meeting_intelligence_prompt(
-            request,
-            prompt_template,
-        )
-        if not self.model:
-            raise ValueError(
-                "No OpenAI model configured. Please select a model in Settings."
-            )
-        request_kwargs = {
-            "model": self.model,
-            "messages": [{"role": "user", "content": prompt}],
-            "timeout": timeout,
-        }
-        if self.model.startswith("gpt") or "gpt" in self.model:
-            request_kwargs["temperature"] = 0.2
-        try:
-            response = self.client.chat.completions.create(**request_kwargs)
-            text = response.choices[0].message.content or ""
-            return self.parse_automatic_meeting_intelligence_result(text, request)
-        except Exception as e:  # noqa: BLE001
-            if "not a chat model" in str(e) or "404" in str(e):
-                logger.error(
-                    f"OpenAI API error (meeting intelligence): Invalid model {self.model}. {e}"
-                )
-                raise ValueError(
-                    f"The model '{self.model}' appears to be invalid or is not a chat model. Please check the model name in Settings."
-                )
-            logger.error(f"OpenAI API error (meeting intelligence): {e}")
-            raise RuntimeError(f"OpenAI API error (meeting intelligence): {e}")
-
-    def generate_meeting_edge(
-        self,
-        request: MeetingEdgeRequest,
-        prompt_template: str = None,
-        timeout: int = 60,
-    ) -> MeetingEdgeResult:
-        prompt = self.build_meeting_edge_prompt(request, prompt_template)
-        if not self.model:
-            raise ValueError(
-                "No OpenAI model configured. Please select a model in Settings."
-            )
-
-        request_kwargs = {
-            "model": self.model,
-            "messages": [{"role": "user", "content": prompt}],
-            "timeout": timeout,
-        }
-        if self.model.startswith("gpt") or "gpt" in self.model:
-            request_kwargs["temperature"] = 0.2
-
-        try:
-            try:
-                response = self.client.chat.completions.create(
-                    **request_kwargs,
-                    response_format={"type": "json_object"},
-                )
-            except Exception as json_mode_error:  # noqa: BLE001
-                # OpenAI-compatible endpoints may reject response_format; retry without it.
-                logger.warning(
-                    "OpenAI JSON mode failed for Meeting Edge (%s); retrying without response_format.",
-                    json_mode_error,
-                )
-                response = self.client.chat.completions.create(**request_kwargs)
-            text = response.choices[0].message.content or ""
-            return self.parse_meeting_edge_result(text, request)
-        except Exception as e:  # noqa: BLE001
-            if "not a chat model" in str(e) or "404" in str(e):
-                logger.error(
-                    f"OpenAI API error (Meeting Edge): Invalid model {self.model}. {e}"
-                )
-                raise ValueError(
-                    f"The model '{self.model}' appears to be invalid or is not a chat model. Please check the model name in Settings."
-                )
-            logger.error(f"OpenAI API error (Meeting Edge): {e}")
-            raise RuntimeError(f"OpenAI API error (Meeting Edge): {e}")
-
-    # infer_speakers_and_generate_notes is inherited and calls the above two methods
-
-    def ask_question_about_meeting(
-        self,
-        user_question: str,
-        meeting_notes: str,
-        diarized_transcript: str,
-        conversation_history: list = None,
-        timeout: int = 60,
-        recording_id: str = None,
-    ):
-        if recording_id is not None:
-            diarized_transcript = self.get_mapped_transcript_for_llm(recording_id)
-
-        # Lead with the stable meeting context as a system message so OpenAI's
-        # automatic prefix caching can reuse it across turns; the volatile
-        # question is sent last.
-        context = build_chat_context(meeting_notes, diarized_transcript)
-        messages = [{"role": "system", "content": context}] + build_chat_messages(
-            user_question, conversation_history
-        )
-        if not self.model:
-            raise ValueError(
-                "No OpenAI model configured. Please select a model in Settings."
-            )
-        try:
-            response = self.client.chat.completions.create(
-                model=self.model, messages=messages, temperature=0.2, timeout=timeout
-            )
-            return response.choices[0].message.content
-        except Exception as e:  # noqa: BLE001
-            if "not a chat model" in str(e) or "404" in str(e):
-                logger.error(
-                    f"OpenAI API error (chat): Invalid model {self.model}. {e}"
-                )
-                raise ValueError(
-                    f"The model '{self.model}' appears to be invalid or is not a chat model. Please check the model name in Settings."
-                )
-            logger.error(f"OpenAI API error (chat): {e}")
-            raise RuntimeError(f"OpenAI API error (chat): {e}")
-
-    def ask_question_streaming(
-        self,
-        user_question: str,
-        meeting_notes: str,
-        diarized_transcript: str,
-        conversation_history: list = None,
-        timeout: int = 60,
-        recording_id: str = None,
-    ) -> Generator[Any, None, None]:
-        if recording_id is not None:
-            diarized_transcript = self.get_mapped_transcript_for_llm(recording_id)
-
-        # Lead with the stable meeting context as a system message so OpenAI's
-        # automatic prefix caching can reuse it across turns; the volatile
-        # question is sent last.
-        context = build_chat_context(meeting_notes, diarized_transcript)
-        messages = [{"role": "system", "content": context}] + build_chat_messages(
-            user_question, conversation_history
-        )
-
-        tools = [
-            {
-                "type": "function",
-                "function": {
-                    "name": "update_meeting_notes",
-                    "description": "Overwrites the current meeting notes with new content. Use this whenever the user asks to modify, edit, add to, or delete parts of the meeting notes. The input should be the fully updated Markdown content of the notes.",
-                    "parameters": {
-                        "type": "object",
-                        "properties": {
-                            "content": {
-                                "type": "string",
-                                "description": "The full, updated Markdown text of the meeting notes.",
-                            }
-                        },
-                        "required": ["content"],
-                    },
-                },
-            }
-        ]
-
-        # Prepare request arguments
-        request_kwargs = {
-            "model": self.model,
-            "messages": messages,
-            "stream": True,
-            "timeout": timeout,
-            "tools": tools,
-        }
-
-        # Skip temperature for reasoning models (OpenAI) that enforce default temperature
-        if self.model.startswith("gpt") or "gpt" in self.model:
-            request_kwargs["temperature"] = 0.2
-
-        try:
-            stream = self.client.chat.completions.create(**request_kwargs)
-
-            tool_calls_accumulator = {}
-
-            for chunk in stream:
-                if not chunk.choices:
-                    continue
-
-                delta = chunk.choices[0].delta
-
-                if delta.content is not None:
-                    yield delta.content
-
-                if delta.tool_calls:
-                    for tool_part in delta.tool_calls:
-                        idx = tool_part.index
-                        if idx not in tool_calls_accumulator:
-                            tool_calls_accumulator[idx] = {"name": "", "arguments": ""}
-
-                        if tool_part.function and tool_part.function.name:
-                            tool_calls_accumulator[idx]["name"] += (
-                                tool_part.function.name
-                            )
-
-                        if tool_part.function and tool_part.function.arguments:
-                            tool_calls_accumulator[idx]["arguments"] += (
-                                tool_part.function.arguments
-                            )
-
-            # Process accumulated tool calls
-            for idx, tool_data in tool_calls_accumulator.items():
-                if tool_data["name"] == "update_meeting_notes":
-                    try:
-                        args = json.loads(tool_data["arguments"])
-                        new_notes = args.get("content")
-                        if new_notes and recording_id:
-                            self._update_notes_in_db(recording_id, new_notes)
-                            yield {"type": "notes_update"}
-                            yield "I have updated the meeting notes."
-                    except json.JSONDecodeError:
-                        logger.error(
-                            "Failed to parse tool arguments for update_meeting_notes"
-                        )
-
-        except Exception as e:  # noqa: BLE001
-            if "not a chat model" in str(e) or "404" in str(e):
-                logger.error(
-                    f"OpenAI API error (streaming chat): Invalid model {self.model}. {e}"
-                )
-                raise ValueError(
-                    f"The model '{self.model}' appears to be invalid or is not a chat model. Please check the model name in Settings."
-                )
-            logger.error(f"OpenAI API error (streaming chat): {e}")
-            raise RuntimeError(f"OpenAI API error (streaming chat): {e}")
-
-    def infer_meeting_title(
-        self,
-        transcript: str,
-        prompt_template: str = None,
-        timeout: int = 60,
-        output_language_instruction: Optional[str] = None,
-    ) -> str:
-        """
-        Infer a concise, descriptive meeting title from the provided transcript.
-        Sub-classes must implement.
-        """
-        if prompt_template is None:
-            prompt_template = self.get_title_prompt_template()
-        prompt = prompt_template.format(
-            transcript=transcript,
-            output_language_section=build_output_language_prompt_section(
-                output_language_instruction
-            ),
-        )
-        if not self.model:
-            raise ValueError(
-                "No OpenAI model configured. Please select a model in Settings."
-            )
-        try:
-            response = self.client.chat.completions.create(
-                model=self.model,
-                messages=[{"role": "user", "content": prompt}],
-                temperature=0.2,
-                timeout=timeout,
-            )
-            title = self.parse_title(response.choices[0].message.content)
-            return title
-        except Exception as e:  # noqa: BLE001
-            if "not a chat model" in str(e) or "404" in str(e):
-                logger.error(
-                    f"OpenAI API error (meeting title): Invalid model {self.model}. {e}"
-                )
-                raise ValueError(
-                    f"The model '{self.model}' appears to be invalid or is not a chat model. Please check the model name in Settings."
-                )
-            logger.error(f"OpenAI API error (meeting title): {e}")
-            raise RuntimeError(f"OpenAI API error (meeting title): {e}")
-
-    def validate_api_key(self) -> bool:
-        """
-        Validate the API key by making a lightweight API call.
-        Returns True if valid, raises an exception or returns False if invalid.
-        """
-        try:
-            # Simple call to list models to verify key
-            self.client.models.list()
-            return True
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"OpenAI API validation failed: {e}")
-            raise ValueError(f"OpenAI API validation failed: {e}")
-
-
-class AnthropicLLMBackend(LLMBackend):
-    def __init__(self, api_key=None, model=None):
-        import anthropic
-
-        if api_key is None:
-            api_key = config_manager.get("anthropic_api_key")
-        if not api_key:
-            raise ValueError(
-                "Anthropic API key is not set. Please provide it in settings."
-            )
-        self.api_key = api_key
-        self.model = model or _get_default_model_for_provider("anthropic")
-        self.client = anthropic.Anthropic(api_key=self.api_key)
-
-    def list_models(self) -> List[str]:
-        """
-        List available Anthropic models.
-        """
-        try:
-            # Anthropic Python SDK supports models.list()
-            if hasattr(self.client, "models") and hasattr(self.client.models, "list"):
-                models = self.client.models.list()
-                # Filter for claude models
-                return sorted([m.id for m in models if "claude" in m.id])
-            else:
-                return []
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Anthropic API error (list models): {e}")
-            return []
-
-    def infer_speaker_suggestions(
-        self,
-        transcript: str,
-        prompt_template: str = None,
-        timeout: int = 60,
-        user_notes: Optional[str] = None,
-        meeting_context: Optional[MeetingEventContext] = None,
-        eligible_labels: Optional[Sequence[str]] = None,
-    ) -> SpeakerInferenceResult:
-        """
-        Run speaker inference on the transcript and return structured suggestions.
-        Can be called independently of meeting notes generation.
-        """
-        if prompt_template is None:
-            prompt_template = self.get_speaker_suggestion_prompt_template()
-        prompt = self.build_speaker_suggestion_prompt(
-            prompt_template,
-            transcript,
-            eligible_labels,
-            user_notes,
-            meeting_context,
-        )
-        if not self.model:
-            raise ValueError(
-                "No Anthropic model configured. Please select a model in Settings."
-            )
-        try:
-            response = self.client.messages.create(
-                model=self.model,
-                max_tokens=1024,
-                messages=[{"role": "user", "content": prompt}],
-                temperature=0.2,
-            )
-            text = (
-                response.content[0].text
-                if hasattr(response.content[0], "text")
-                else response.content[0]
-            )
-            return self.parse_speaker_inference_result(text, eligible_labels)
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Anthropic API error (speaker suggestions): {e}")
-            raise RuntimeError(f"Anthropic API error (speaker suggestions): {e}")
-
-    def infer_speakers(
-        self,
-        transcript: str,
-        prompt_template: str = None,
-        timeout: int = 60,
-        user_notes: Optional[str] = None,
-        meeting_context: Optional[MeetingEventContext] = None,
-        eligible_labels: Optional[Sequence[str]] = None,
-    ) -> Dict[str, str]:
-        return self.infer_speaker_suggestions(
-            transcript,
-            prompt_template,
-            timeout,
-            user_notes=user_notes,
-            meeting_context=meeting_context,
-            eligible_labels=eligible_labels,
-        ).mapping
-
-    def generate_meeting_notes(
-        self,
-        transcript: str,
-        speaker_mapping: Dict[str, str],
-        prompt_template: str = None,
-        timeout: int = 60,
-        user_notes: Optional[str] = None,
-        meeting_context: Optional[MeetingEventContext] = None,
-        output_language_instruction: Optional[str] = None,
-    ) -> str:
-        """
-        Generate meeting notes using the provided speaker mapping. Should be called after user relabeling.
-        """
-        if prompt_template is None:
-            prompt_template = self.get_notes_prompt_template()
-        prompt = self.build_notes_prompt(
-            prompt_template,
-            transcript,
-            speaker_mapping,
-            user_notes,
-            meeting_context,
-            output_language_instruction,
-        )
-        if not self.model:
-            raise ValueError(
-                "No Anthropic model configured. Please select a model in Settings."
-            )
-        try:
-            response = self.client.messages.create(
-                model=self.model,
-                max_tokens=2048,
-                messages=[{"role": "user", "content": prompt}],
-                temperature=0.2,
-            )
-            text = (
-                response.content[0].text
-                if hasattr(response.content[0], "text")
-                else response.content[0]
-            )
-            notes = self.finalise_meeting_notes(self.parse_notes(text), user_notes)
-            return notes
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Anthropic API error (meeting notes): {e}")
-            raise RuntimeError(f"Anthropic API error (meeting notes): {e}")
-
-    def generate_meeting_intelligence(
-        self,
-        request: AutomaticMeetingIntelligenceRequest,
-        prompt_template: str = None,
-        timeout: int = 60,
-    ) -> AutomaticMeetingIntelligenceResult:
-        prompt = self.build_automatic_meeting_intelligence_prompt(
-            request,
-            prompt_template,
-        )
-        if not self.model:
-            raise ValueError(
-                "No Anthropic model configured. Please select a model in Settings."
-            )
-        try:
-            response = self.client.messages.create(
-                model=self.model,
-                max_tokens=4096,
-                messages=[{"role": "user", "content": prompt}],
-                temperature=0.2,
-            )
-            text = (
-                response.content[0].text
-                if hasattr(response.content[0], "text")
-                else response.content[0]
-            )
-            return self.parse_automatic_meeting_intelligence_result(text, request)
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Anthropic API error (meeting intelligence): {e}")
-            raise RuntimeError(f"Anthropic API error (meeting intelligence): {e}")
-
-    def generate_meeting_edge(
-        self,
-        request: MeetingEdgeRequest,
-        prompt_template: str = None,
-        timeout: int = 60,
-    ) -> MeetingEdgeResult:
-        # Split into a stable instruction prefix and the volatile per-refresh
-        # context so the prefix can be cached across refreshes. The parts
-        # concatenate to the same prompt, so the model output is unchanged.
-        prefix, suffix = self.build_meeting_edge_prompt_parts(request, prompt_template)
-        if prefix:
-            user_content = [
-                {
-                    "type": "text",
-                    "text": prefix,
-                    "cache_control": {"type": "ephemeral"},
-                },
-                {"type": "text", "text": suffix},
-            ]
-        else:
-            user_content = suffix
-        if not self.model:
-            raise ValueError(
-                "No Anthropic model configured. Please select a model in Settings."
-            )
-        try:
-            # No assistant prefill: a last-assistant-turn prefill 400s on current
-            # Claude models (Opus 4.6+, Sonnet 5, Fable 5). The prompt already
-            # mandates a JSON object and the tolerant parser handles any
-            # fenced/prose wrapping, so raw JSON is not forced by a prefill.
-            response = self.client.messages.create(
-                model=self.model,
-                max_tokens=2048,
-                messages=[
-                    {"role": "user", "content": user_content},
-                ],
-                temperature=0.2,
-                timeout=timeout,
-            )
-            text = (
-                response.content[0].text
-                if hasattr(response.content[0], "text")
-                else response.content[0]
-            )
-            return self.parse_meeting_edge_result(text, request)
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Anthropic API error (Meeting Edge): {e}")
-            raise RuntimeError(f"Anthropic API error (Meeting Edge): {e}")
-
-    # infer_speakers_and_generate_notes is inherited and calls the above two methods
-
-    def ask_question_about_meeting(
-        self,
-        user_question: str,
-        meeting_notes: str,
-        diarized_transcript: str,
-        conversation_history: list = None,
-        timeout: int = 60,
-        recording_id: str = None,
-    ):
-        # If recording_id is provided, use mapped transcript
-        if recording_id is not None:
-            diarized_transcript = self.get_mapped_transcript_for_llm(recording_id)
-
-        # Put the stable meeting context in the cached system prompt (render order
-        # is tools -> system -> messages), so it is reused across turns while the
-        # messages array carries only the volatile history and question.
-        context = build_chat_context(meeting_notes, diarized_transcript)
-        messages = build_chat_messages(user_question, conversation_history)
-        if not self.model:
-            raise ValueError(
-                "No Anthropic model configured. Please select a model in Settings."
-            )
-        try:
-            response = self.client.messages.create(
-                model=self.model,
-                max_tokens=1024,
-                system=anthropic_cached_system(context),
-                messages=messages,
-                temperature=0.2,
-            )
-            return (
-                response.content[0].text
-                if hasattr(response.content[0], "text")
-                else response.content[0]
-            )
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Anthropic API error (chat): {e}")
-            raise RuntimeError(f"Anthropic API error (chat): {e}")
-
-    def ask_question_streaming(
-        self,
-        user_question: str,
-        meeting_notes: str,
-        diarized_transcript: str,
-        conversation_history: list = None,
-        timeout: int = 60,
-        recording_id: str = None,
-    ) -> Generator[Any, None, None]:
-        if recording_id is not None:
-            diarized_transcript = self.get_mapped_transcript_for_llm(recording_id)
-
-        # Put the stable meeting context in the cached system prompt (render order
-        # is tools -> system -> messages), so it is reused across turns while the
-        # messages array carries only the volatile history and question.
-        context = build_chat_context(meeting_notes, diarized_transcript)
-        messages = build_chat_messages(user_question, conversation_history)
-
-        tool_definition = {
-            "name": "update_meeting_notes",
-            "description": "Overwrites the current meeting notes with new content. Use this whenever the user asks to modify, edit, add to, or delete parts of the meeting notes. The input should be the fully updated Markdown content of the notes.",
-            "input_schema": {
-                "type": "object",
-                "properties": {
-                    "content": {
-                        "type": "string",
-                        "description": "The full, updated Markdown text of the meeting notes.",
-                    }
-                },
-                "required": ["content"],
-            },
-        }
-
-        try:
-            with self.client.messages.stream(
-                model=self.model,
-                max_tokens=1024,
-                system=anthropic_cached_system(context),
-                messages=messages,
-                temperature=0.2,
-                tools=[tool_definition],
-            ) as stream:
-                current_tool_name = None
-                current_json_accum = ""
-
-                for event in stream:
-                    # Text Delta
-                    if (
-                        event.type == "content_block_delta"
-                        and event.delta.type == "text_delta"
-                    ):
-                        yield event.delta.text
-
-                    # Tool Start
-                    elif (
-                        event.type == "content_block_start"
-                        and event.content_block.type == "tool_use"
-                    ):
-                        current_tool_name = event.content_block.name
-                        current_json_accum = ""
-
-                    # Tool Args Delta
-                    elif (
-                        event.type == "content_block_delta"
-                        and event.delta.type == "input_json_delta"
-                    ):
-                        current_json_accum += event.delta.partial_json
-
-                    # Tool Stop (Execute)
-                    elif event.type == "content_block_stop":
-                        if current_tool_name == "update_meeting_notes":
-                            try:
-                                args = json.loads(current_json_accum)
-                                new_notes = args.get("content")
-                                if new_notes and recording_id:
-                                    self._update_notes_in_db(recording_id, new_notes)
-                                    yield {"type": "notes_update"}
-                                    yield "I have updated the meeting notes."
-                            except json.JSONDecodeError:
-                                logger.error(
-                                    "Failed to parse tool arguments for update_meeting_notes"
-                                )
-                            finally:
-                                current_tool_name = None
-
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Anthropic API error (streaming chat): {e}")
-            raise RuntimeError(f"Anthropic API error (streaming chat): {e}")
-
-    def infer_meeting_title(
-        self,
-        transcript: str,
-        prompt_template: str = None,
-        timeout: int = 60,
-        output_language_instruction: Optional[str] = None,
-    ) -> str:
-        """
-        Infer a concise, descriptive meeting title from the provided transcript.
-        Sub-classes must implement.
-        """
-        if prompt_template is None:
-            prompt_template = self.get_title_prompt_template()
-        prompt = prompt_template.format(
-            transcript=transcript,
-            output_language_section=build_output_language_prompt_section(
-                output_language_instruction
-            ),
-        )
-        try:
-            response = self.client.messages.create(
-                model=self.model,
-                max_tokens=1024,
-                messages=[{"role": "user", "content": prompt}],
-                temperature=0.2,
-            )
-            title = self.parse_title(
-                response.content[0].text
-                if hasattr(response.content[0], "text")
-                else response.content[0]
-            )
-            return title
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Anthropic API error (meeting title): {e}")
-            raise RuntimeError(f"Anthropic API error (meeting title): {e}")
-
-    def validate_api_key(self) -> bool:
-        """
-        Validate the API key by making a lightweight API call.
-        Returns True if valid, raises an exception or returns False if invalid.
-        """
-        try:
-            # Use list_models to verify key without consuming tokens
-            if hasattr(self.client, "models") and hasattr(self.client.models, "list"):
-                self.client.models.list()
-                return True
-
-            # Fallback if list_models not available (should not happen with recent SDK)
-            if not self.model:
-                raise ValueError(
-                    "No Anthropic model configured. Please select a model in Settings."
-                )
-
-            # Try a minimal generation with the configured model
-            self.client.messages.create(
-                model=self.model,
-                max_tokens=1,
-                messages=[{"role": "user", "content": "Hi"}],
-            )
-            return True
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Anthropic API validation failed: {e}")
-            raise ValueError(f"Anthropic API validation failed: {e}")
-
-
-# Ollama's num_ctx defaults to 2048, which silently truncates meeting-length prompts.
-OLLAMA_DEFAULT_NUM_CTX = 8192
-
-
-class OllamaLLMBackend(LLMBackend):
-    def __init__(
-        self,
-        api_url=None,
-        model=None,
-        context_window: int | None = None,
-        allow_private_api_url: bool = False,
-    ):
-        import requests
-
-        self.requests = requests
-        trusted_api_url = config_manager.get("ollama_api_url")
-        if api_url is None:
-            api_url = trusted_api_url
-        if not api_url:
-            api_url = "http://host.docker.internal:11434"
-
-        self.api_url = validate_ollama_api_url(
-            api_url,
-            allow_private=allow_private_api_url,
-            trusted_url=trusted_api_url,
-        )
-        self.model = model or config_manager.get("ollama_model")
-        self.context_window = context_window or config_manager.get(
-            "ollama_context_window"
-        )
-
-    def _chat_options(self, *, temperature: float) -> dict[str, object]:
-        # num_ctx must always be sent; an unset one lets Ollama fall back to its
-        # 2048 default, which silently truncates meeting-length prompts.
-        ctx = getattr(self, "context_window", None) or OLLAMA_DEFAULT_NUM_CTX
-        return {"temperature": temperature, "num_ctx": int(ctx)}
-
-    @staticmethod
-    def _raise_if_truncated(response_metadata: dict | None) -> None:
-        if not response_metadata or response_metadata.get("done_reason") != "length":
-            return
-        prompt_eval_count = response_metadata.get("prompt_eval_count")
-        eval_count = response_metadata.get("eval_count")
-        raise RuntimeError(
-            "Ollama stopped because the context window was exhausted "
-            f"(prompt_eval_count={prompt_eval_count}, eval_count={eval_count}). "
-            "Increase the Ollama context window or select a model with a larger context."
-        )
-
-    def _get(self, path: str, **kwargs):
-        return self.requests.get(
-            f"{self.api_url}{path}",
-            allow_redirects=False,
-            **kwargs,
-        )
-
-    def _post(self, path: str, **kwargs):
-        return self.requests.post(
-            f"{self.api_url}{path}",
-            allow_redirects=False,
-            **kwargs,
-        )
-
-    def list_models(self) -> List[str]:
-        try:
-            resp = self._get("/api/tags", timeout=10)
-            resp.raise_for_status()
-            data = resp.json()
-            return sorted([m["name"] for m in data.get("models", [])])
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Ollama API error (list models): {e}")
-            return []
-
-    def infer_speaker_suggestions(
-        self,
-        transcript: str,
-        prompt_template: str = None,
-        timeout: int = 60,
-        user_notes: Optional[str] = None,
-        meeting_context: Optional[MeetingEventContext] = None,
-        eligible_labels: Optional[Sequence[str]] = None,
-    ) -> SpeakerInferenceResult:
-        if prompt_template is None:
-            prompt_template = self.get_speaker_suggestion_prompt_template()
-        prompt = self.build_speaker_suggestion_prompt(
-            prompt_template,
-            transcript,
-            eligible_labels,
-            user_notes,
-            meeting_context,
-        )
-        if not self.model:
-            raise ValueError(
-                "No Ollama model configured. Please select a model in Settings."
-            )
-
-        try:
-            payload = {
-                "model": self.model,
-                "messages": [{"role": "user", "content": prompt}],
-                "stream": False,
-                "options": self._chat_options(temperature=0.3),
-            }
-            resp = self._post("/api/chat", json=payload, timeout=timeout)
-            resp.raise_for_status()
-            response_json = resp.json()
-            self._raise_if_truncated(response_json)
-            text = response_json.get("message", {}).get("content", "")
-            return self.parse_speaker_inference_result(text, eligible_labels)
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Ollama API error (speaker suggestions): {e}")
-            raise RuntimeError(f"Ollama API error (speaker suggestions): {e}")
-
-    def infer_speakers(
-        self,
-        transcript: str,
-        prompt_template: str = None,
-        timeout: int = 60,
-        user_notes: Optional[str] = None,
-        meeting_context: Optional[MeetingEventContext] = None,
-        eligible_labels: Optional[Sequence[str]] = None,
-    ) -> Dict[str, str]:
-        return self.infer_speaker_suggestions(
-            transcript,
-            prompt_template,
-            timeout,
-            user_notes=user_notes,
-            meeting_context=meeting_context,
-            eligible_labels=eligible_labels,
-        ).mapping
-
-    def generate_meeting_notes(
-        self,
-        transcript: str,
-        speaker_mapping: Dict[str, str],
-        prompt_template: str = None,
-        timeout: int = 60,
-        user_notes: Optional[str] = None,
-        meeting_context: Optional[MeetingEventContext] = None,
-        output_language_instruction: Optional[str] = None,
-    ) -> str:
-        if prompt_template is None:
-            prompt_template = self.get_notes_prompt_template()
-        prompt = self.build_notes_prompt(
-            prompt_template,
-            transcript,
-            speaker_mapping,
-            user_notes,
-            meeting_context,
-            output_language_instruction,
-        )
-        if not self.model:
-            raise ValueError(
-                "No Ollama model configured. Please select a model in Settings."
-            )
-
-        try:
-            payload = {
-                "model": self.model,
-                "messages": [{"role": "user", "content": prompt}],
-                "stream": False,
-                "options": self._chat_options(temperature=0.3),
-            }
-            resp = self._post("/api/chat", json=payload, timeout=timeout)
-            resp.raise_for_status()
-            response_json = resp.json()
-            self._raise_if_truncated(response_json)
-            text = response_json.get("message", {}).get("content", "")
-            return self.finalise_meeting_notes(self.parse_notes(text), user_notes)
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Ollama API error (meeting notes): {e}")
-            raise RuntimeError(f"Ollama API error (meeting notes): {e}")
-
-    def generate_meeting_intelligence(
-        self,
-        request: AutomaticMeetingIntelligenceRequest,
-        prompt_template: str = None,
-        timeout: int = 60,
-    ) -> AutomaticMeetingIntelligenceResult:
-        prompt = self.build_automatic_meeting_intelligence_prompt(
-            request,
-            prompt_template,
-        )
-        if not self.model:
-            raise ValueError(
-                "No Ollama model configured. Please select a model in Settings."
-            )
-
-        try:
-            payload = {
-                "model": self.model,
-                "messages": [{"role": "user", "content": prompt}],
-                "stream": False,
-                "format": "json",
-                "options": self._chat_options(temperature=0.3),
-            }
-            resp = self._post("/api/chat", json=payload, timeout=timeout)
-            resp.raise_for_status()
-            response_json = resp.json()
-            self._raise_if_truncated(response_json)
-            text = response_json.get("message", {}).get("content", "")
-            try:
-                return self.parse_automatic_meeting_intelligence_result(text, request)
-            except JSON_CONTRACT_ERRORS as parse_error:
-                logger.warning(
-                    "Ollama meeting intelligence response failed JSON contract; retrying repair: %s; response_shape=%s",
-                    parse_error,
-                    summarize_llm_response_shape(text),
-                )
-                repair_payload = {
-                    "model": self.model,
-                    "messages": [
-                        {
-                            "role": "user",
-                            "content": self.build_json_repair_prompt(
-                                original_prompt=prompt,
-                                invalid_response=text,
-                                validation_error=parse_error,
-                            ),
-                        }
-                    ],
-                    "stream": False,
-                    "format": "json",
-                    "options": self._chat_options(temperature=0.0),
-                }
-                repair_resp = self._post(
-                    "/api/chat", json=repair_payload, timeout=timeout
-                )
-                repair_resp.raise_for_status()
-                repair_json = repair_resp.json()
-                self._raise_if_truncated(repair_json)
-                repair_text = repair_json.get("message", {}).get("content", "")
-                return self.parse_automatic_meeting_intelligence_result(
-                    repair_text, request
-                )
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Ollama API error (meeting intelligence): {e}")
-            raise RuntimeError(f"Ollama API error (meeting intelligence): {e}")
-
-    def generate_meeting_edge(
-        self,
-        request: MeetingEdgeRequest,
-        prompt_template: str = None,
-        timeout: int = 60,
-    ) -> MeetingEdgeResult:
-        prompt = self.build_meeting_edge_prompt(request, prompt_template)
-        if not self.model:
-            raise ValueError(
-                "No Ollama model configured. Please select a model in Settings."
-            )
-
-        try:
-            payload = {
-                "model": self.model,
-                "messages": [{"role": "user", "content": prompt}],
-                "stream": False,
-                "format": "json",
-                "options": self._chat_options(temperature=0.3),
-            }
-            resp = self._post("/api/chat", json=payload, timeout=timeout)
-            resp.raise_for_status()
-            response_json = resp.json()
-            self._raise_if_truncated(response_json)
-            text = response_json.get("message", {}).get("content", "")
-            try:
-                return self.parse_meeting_edge_result(text, request)
-            except JSON_CONTRACT_ERRORS as parse_error:
-                logger.warning(
-                    "Ollama Meeting Edge response failed JSON contract; retrying repair: %s; response_shape=%s",
-                    parse_error,
-                    summarize_llm_response_shape(text),
-                )
-                repair_payload = {
-                    "model": self.model,
-                    "messages": [
-                        {
-                            "role": "user",
-                            "content": self.build_json_repair_prompt(
-                                original_prompt=prompt,
-                                invalid_response=text,
-                                validation_error=parse_error,
-                            ),
-                        }
-                    ],
-                    "stream": False,
-                    "format": "json",
-                    "options": self._chat_options(temperature=0.0),
-                }
-                repair_resp = self._post(
-                    "/api/chat", json=repair_payload, timeout=timeout
-                )
-                repair_resp.raise_for_status()
-                repair_json = repair_resp.json()
-                self._raise_if_truncated(repair_json)
-                repair_text = repair_json.get("message", {}).get("content", "")
-                return self.parse_meeting_edge_result(repair_text, request)
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Ollama API error (Meeting Edge): {e}")
-            raise RuntimeError(f"Ollama API error (Meeting Edge): {e}")
-
-    def ask_question_about_meeting(
-        self,
-        user_question: str,
-        meeting_notes: str,
-        diarized_transcript: str,
-        conversation_history: list = None,
-        timeout: int = 60,
-        recording_id: str = None,
-    ):
-        if recording_id is not None:
-            diarized_transcript = self.get_mapped_transcript_for_llm(recording_id)
-
-        prompt = self._build_chat_prompt(
-            user_question, meeting_notes, diarized_transcript
-        )
-
-        messages = []
-        if conversation_history:
-            for msg in conversation_history:
-                if msg.get("role") and msg.get("parts"):
-                    # Ollama's /api/chat accepts only user/assistant/system, so
-                    # normalise the Gemini-style 'model' history role to 'assistant'.
-                    role = "assistant" if msg["role"] == "model" else msg["role"]
-                    for part in msg["parts"]:
-                        messages.append({"role": role, "content": part["text"]})
-        messages.append({"role": "user", "content": prompt})
-
-        if not self.model:
-            raise ValueError(
-                "No Ollama model configured. Please select a model in Settings."
-            )
-
-        try:
-            payload = {
-                "model": self.model,
-                "messages": messages,
-                "stream": False,
-                "options": self._chat_options(temperature=0.3),
-            }
-            resp = self._post("/api/chat", json=payload, timeout=timeout)
-            resp.raise_for_status()
-            response_json = resp.json()
-            self._raise_if_truncated(response_json)
-            return response_json.get("message", {}).get("content", "")
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Ollama API error (chat): {e}")
-            raise RuntimeError(f"Ollama API error (chat): {e}")
-
-    def ask_question_streaming(
-        self,
-        user_question: str,
-        meeting_notes: str,
-        diarized_transcript: str,
-        conversation_history: list = None,
-        timeout: int = 60,
-        recording_id: str = None,
-    ) -> Generator[str, None, None]:
-        if recording_id is not None:
-            diarized_transcript = self.get_mapped_transcript_for_llm(recording_id)
-
-        prompt = self._build_chat_prompt(
-            user_question, meeting_notes, diarized_transcript
-        )
-
-        messages = []
-        if conversation_history:
-            for msg in conversation_history:
-                if msg.get("role") and msg.get("parts"):
-                    # Ollama's /api/chat accepts only user/assistant/system, so
-                    # normalise the Gemini-style 'model' history role to 'assistant'.
-                    role = "assistant" if msg["role"] == "model" else msg["role"]
-                    for part in msg["parts"]:
-                        messages.append({"role": role, "content": part["text"]})
-        messages.append({"role": "user", "content": prompt})
-
-        if not self.model:
-            raise ValueError(
-                "No Ollama model configured. Please select a model in Settings."
-            )
-
-        try:
-            payload = {
-                "model": self.model,
-                "messages": messages,
-                "stream": True,
-                "options": self._chat_options(temperature=0.3),
-            }
-            resp = self._post("/api/chat", json=payload, stream=True, timeout=timeout)
-            resp.raise_for_status()
-
-            import json
-
-            final_metadata = None
-            for line in resp.iter_lines():
-                if line:
-                    try:
-                        chunk = json.loads(line)
-                        if chunk.get("done"):
-                            final_metadata = chunk
-                        content = chunk.get("message", {}).get("content", "")
-                        if content:
-                            yield content
-                    except json.JSONDecodeError:
-                        pass
-            self._raise_if_truncated(final_metadata)
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Ollama API error (streaming chat): {e}")
-            raise RuntimeError(f"Ollama API error (streaming chat): {e}")
-
-    def infer_meeting_title(
-        self,
-        transcript: str,
-        prompt_template: str = None,
-        timeout: int = 60,
-        output_language_instruction: Optional[str] = None,
-    ) -> str:
-        if prompt_template is None:
-            prompt_template = self.get_title_prompt_template()
-        prompt = prompt_template.format(
-            transcript=transcript,
-            output_language_section=build_output_language_prompt_section(
-                output_language_instruction
-            ),
-        )
-        if not self.model:
-            raise ValueError(
-                "No Ollama model configured. Please select a model in Settings."
-            )
-        try:
-            payload = {
-                "model": self.model,
-                "messages": [{"role": "user", "content": prompt}],
-                "stream": False,
-                "options": self._chat_options(temperature=0.3),
-            }
-            resp = self._post("/api/chat", json=payload, timeout=timeout)
-            resp.raise_for_status()
-            response_json = resp.json()
-            self._raise_if_truncated(response_json)
-            text = response_json.get("message", {}).get("content", "")
-            return self.parse_title(text)
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Ollama API error (meeting title): {e}")
-            raise RuntimeError(f"Ollama API error (meeting title): {e}")
-
-    def validate_api_key(self) -> bool:
-        try:
-            self.list_models()
-            return True
-        except Exception as e:  # noqa: BLE001
-            logger.error(f"Ollama API validation failed: {e}")
-            raise ValueError(f"Ollama API validation failed: {e}")
-
-
-def _get_default_model_for_provider(provider: str) -> Optional[str]:
-    """Return the recommended default model for a provider."""
-    # Defaults are no longer hardcoded here.
-    # Users must select a model or rely on frontend recommendations.
-    return None
-
-
-def get_default_model_for_provider(provider: str) -> Optional[str]:
-    """
-    Public function to get the default model for a provider.
-    This is the single source of truth for default model names.
-    """
-    return _get_default_model_for_provider(provider)
-
-
-# --- LLM Backend Factory ---
-def get_llm_backend(
-    provider: str,
-    api_key=None,
-    model=None,
-    api_url=None,
-    context_window: int | None = None,
-    allow_private_api_url: bool = False,
-    cli_user_id: int | None = None,
-    cli_provider: str | None = None,
-):
-    """
-    Factory function to instantiate the appropriate LLM backend.
-    Heavy dependencies are only imported when needed.
-    Args:
-        provider (str): 'gemini', 'openai', 'anthropic', or 'ollama'
-        api_key (str): API key for the provider (optional)
-        model (str): Model name (optional)
-        api_url (str): API URL for local providers (optional)
-    Returns:
-        Instance of the appropriate LLMBackend subclass.
-    Raises:
-        ValueError: If provider is unknown.
-    """
-    from backend.utils.config_manager import config_manager
-
-    if model is None:
-        if provider == "ollama":
-            model = config_manager.get("ollama_model")
-        else:
-            model = config_manager.get(f"{provider}_model")
-
-    logger.info(
-        "Creating LLM backend: provider=%s model=%s api_url=%s",
-        provider,
-        model,
-        api_url,
-    )
-
-    if provider == "gemini":
-        return GeminiLLMBackend(api_key=api_key, model=model)
-    elif provider == "openai":
-        return OpenAILLMBackend(api_key=api_key, model=model)
-    elif provider == "anthropic":
-        return AnthropicLLMBackend(api_key=api_key, model=model)
-    elif provider == "ollama":
-        return OllamaLLMBackend(
-            api_url=api_url,
-            model=model,
-            context_window=context_window,
-            allow_private_api_url=allow_private_api_url,
-        )
-    elif provider == "cli":
-        from backend.processing.cli_backend import CliLLMBackend
-
-        cli = cli_provider or "claude_code"  # subscription CLI: Claude or Codex
-        return CliLLMBackend(model=model, user_id=cli_user_id, provider=cli)
-    else:
-        raise ValueError(f"Unknown LLM provider: {provider}")
-
-
-class SecondaryLLMBackend(LLMBackend):
-    """Wraps a primary and secondary LLM backend. On any error from primary,
-    falls back to the secondary backend. If both fail, raises the primary error."""
-
-    def __init__(self, primary: LLMBackend, secondary: LLMBackend):
-        self._primary = primary
-        self._secondary = secondary
-        logger.info(
-            "SecondaryLLMBackend: primary=%s secondary=%s",
-            getattr(primary, "model", "unknown"),
-            getattr(secondary, "model", "unknown"),
-        )
-
-    def _call_with_secondary(self, method_name: str, *args, **kwargs):
-        try:
-            result = getattr(self._primary, method_name)(*args, **kwargs)
-            logger.info(
-                "Primary LLM (%s) handled %s successfully.",
-                getattr(self._primary, "model", "unknown"),
-                method_name,
-            )
-            return result
-        except Exception as primary_exc:  # noqa: BLE001 -- boundary: fall back to secondary LLM on any primary failure
-            logger.warning(
-                "Primary LLM (%s) failed on %s: %s. Falling back to secondary (%s).",
-                getattr(self._primary, "model", "unknown"),
-                method_name,
-                primary_exc,
-                getattr(self._secondary, "model", "unknown"),
-            )
-            try:
-                result = getattr(self._secondary, method_name)(*args, **kwargs)
-                logger.info(
-                    "Secondary LLM call succeeded on %s after primary failure.",
-                    method_name,
-                )
-                return result
-            except Exception:  # noqa: BLE001 -- boundary: surface the original primary failure if secondary also fails
-                logger.error("Secondary LLM also failed on %s.", method_name)
-                raise primary_exc
-
-    def _stream_with_secondary(self, method_name: str, *args, **kwargs):
-        has_yielded = False
-        try:
-            gen = getattr(self._primary, method_name)(*args, **kwargs)
-            first_chunk = next(gen)
-            try:
-                yield first_chunk
-                has_yielded = True
-                yield from gen
-            finally:
-                if not has_yielded:
-                    gen.close()
-        except StopIteration:
-            return
-        except Exception as primary_exc:
-            if has_yielded:
-                raise primary_exc
-            logger.warning(
-                "Primary LLM (%s) failed on streaming %s: %s. Falling back to secondary (%s).",
-                getattr(self._primary, "model", "unknown"),
-                method_name,
-                primary_exc,
-                getattr(self._secondary, "model", "unknown"),
-            )
-            try:
-                yield from getattr(self._secondary, method_name)(*args, **kwargs)
-            except Exception:  # noqa: BLE001 -- boundary: surface the original primary failure if secondary also fails
-                raise primary_exc
-
-    def infer_speaker_suggestions(
-        self,
-        transcript,
-        prompt_template=None,
-        timeout=60,
-        user_notes=None,
-        meeting_context=None,
-        eligible_labels=None,
-    ):
-        return self._call_with_secondary(
-            "infer_speaker_suggestions",
-            transcript,
-            prompt_template,
-            timeout,
-            user_notes=user_notes,
-            meeting_context=meeting_context,
-            eligible_labels=eligible_labels,
-        )
-
-    def generate_meeting_intelligence(self, request, prompt_template=None, timeout=60):
-        return self._call_with_secondary(
-            "generate_meeting_intelligence", request, prompt_template, timeout
-        )
-
-    def generate_meeting_edge(self, request, prompt_template=None, timeout=60):
-        return self._call_with_secondary(
-            "generate_meeting_edge", request, prompt_template, timeout
-        )
-
-    def generate_meeting_notes(
-        self,
-        transcript,
-        speaker_mapping,
-        prompt_template=None,
-        timeout=60,
-        user_notes=None,
-        meeting_context=None,
-        output_language_instruction=None,
-    ):
-        return self._call_with_secondary(
-            "generate_meeting_notes",
-            transcript,
-            speaker_mapping,
-            prompt_template,
-            timeout,
-            user_notes=user_notes,
-            meeting_context=meeting_context,
-            output_language_instruction=output_language_instruction,
-        )
-
-    def infer_meeting_title(
-        self,
-        transcript,
-        prompt_template=None,
-        timeout=60,
-        output_language_instruction=None,
-    ):
-        return self._call_with_secondary(
-            "infer_meeting_title",
-            transcript,
-            prompt_template,
-            timeout,
-            output_language_instruction=output_language_instruction,
-        )
-
-    def ask_question_about_meeting(
-        self,
-        user_question,
-        meeting_notes,
-        diarized_transcript,
-        conversation_history=None,
-        timeout=60,
-        recording_id=None,
-    ):
-        return self._call_with_secondary(
-            "ask_question_about_meeting",
-            user_question,
-            meeting_notes,
-            diarized_transcript,
-            conversation_history,
-            timeout,
-            recording_id=recording_id,
-        )
-
-    def ask_question_streaming(
-        self,
-        user_question,
-        meeting_notes,
-        diarized_transcript,
-        conversation_history=None,
-        timeout=60,
-        recording_id=None,
-    ):
-        return self._stream_with_secondary(
-            "ask_question_streaming",
-            user_question,
-            meeting_notes,
-            diarized_transcript,
-            conversation_history,
-            timeout,
-            recording_id=recording_id,
-        )
-
-    def validate_api_key(self) -> bool:
-        try:
-            return self._primary.validate_api_key()
-        except Exception as exc:  # noqa: BLE001 -- boundary: fall back to secondary LLM on any primary failure
-            logger.debug(
-                "Primary LLM validate_api_key failed: %s. Trying secondary.", exc
-            )
-            return self._secondary.validate_api_key()
-
-    def list_models(self) -> list:
-        try:
-            return self._primary.list_models()
-        except Exception as exc:  # noqa: BLE001 -- boundary: fall back to secondary LLM on any primary failure
-            logger.debug("Primary LLM list_models failed: %s. Trying secondary.", exc)
-            return self._secondary.list_models()
-
-
-def get_llm_backend_with_secondary(
-    primary_config,
-    purpose: str = "default",
-):
-    """Instantiate primary LLM backend with optional secondary fallback.
-
-    Args:
-        primary_config: A ResolvedLLMConfig (from resolve_llm_config).
-        purpose: "default" or "meeting_edge".
-
-    Returns:
-        An LLMBackend, wrapped in SecondaryLLMBackend if a secondary
-        provider is configured and operational, otherwise just the primary.
-    """
-    primary = get_llm_backend(
-        provider=primary_config.provider,
-        api_key=primary_config.api_key,
-        model=primary_config.model,
-        api_url=primary_config.api_url,
-        context_window=primary_config.context_window,
-        cli_user_id=primary_config.cli_user_id,
-        cli_provider=primary_config.cli_provider,
-    )
-
-    secondary_cfg = primary_config.secondary_config()
-    if secondary_cfg is None:
-        return primary
-
-    try:
-        # Build the secondary recursively so it can carry its own fallback: a CLI
-        # OAuth config returns the whole server-default chain here, yielding a
-        # nested SecondaryLLMBackend(sub, SecondaryLLMBackend(primary, secondary))
-        # (user sub -> server primary -> server secondary). The ordinary 2-tier
-        # case bottoms out immediately, so its behaviour is unchanged.
-        secondary = get_llm_backend_with_secondary(secondary_cfg, purpose)
-        return SecondaryLLMBackend(primary=primary, secondary=secondary)
-    except Exception as exc:  # noqa: BLE001 -- boundary: continue with primary only if secondary init fails
-        logger.warning(
-            "Failed to initialise secondary LLM backend (%s): %s. "
-            "Continuing with primary only.",
-            secondary_cfg.provider,
-            exc,
-        )
-        return primary
+from backend.processing.llm_backends.anthropic import AnthropicLLMBackend
+from backend.processing.llm_backends.base import (
+    JSON_CONTRACT_ERRORS,
+    LLMBackend,
+    build_eligible_speaker_labels_prompt_section,
+    get_default_model_for_provider,
+    summarize_llm_response_shape,
+)
+from backend.processing.llm_backends.factory import (
+    SecondaryLLMBackend,
+    get_llm_backend,
+    get_llm_backend_with_secondary,
+)
+from backend.processing.llm_backends.gemini import GeminiLLMBackend
+from backend.processing.llm_backends.ollama import (
+    OLLAMA_DEFAULT_NUM_CTX,
+    OllamaLLMBackend,
+)
+from backend.processing.llm_backends.openai import OpenAILLMBackend
+
+# Retained as module attributes for backward compatibility: some call sites and
+# tests reach these through this module (e.g. tests monkeypatch
+# ``llm_services.config_manager``). They are intentionally not part of the
+# public ``__all__`` surface.
+from backend.utils.config_manager import config_manager  # noqa: F401
+from backend.utils.ollama_url_policy import validate_ollama_api_url  # noqa: F401
+
+__all__ = [
+    "JSON_CONTRACT_ERRORS",
+    "OLLAMA_DEFAULT_NUM_CTX",
+    "AnthropicLLMBackend",
+    "GeminiLLMBackend",
+    "LLMBackend",
+    "OllamaLLMBackend",
+    "OpenAILLMBackend",
+    "SecondaryLLMBackend",
+    "build_eligible_speaker_labels_prompt_section",
+    "get_default_model_for_provider",
+    "get_llm_backend",
+    "get_llm_backend_with_secondary",
+    "summarize_llm_response_shape",
+]
