+++ pyproject.toml
@@ -133,7 +133,7 @@ max-complexity = 10
 "backend/processing/engines/onnx_asr_engine.py" = ["C901", "PLR0912"]
 "backend/processing/engines/whisper_engine.py" = ["C901", "PLR0912"]
 "backend/processing/live_transcribe.py" = ["C901", "PLR0912", "PLR0913", "PLR0915"]
-"backend/processing/llm_services.py" = ["C901", "PLR0912", "PLR0913"]
+"backend/processing/llm_backends/*.py" = ["C901", "PLR0912", "PLR0913"]
 "backend/processing/phantom_filter.py" = ["C901", "PLR0912", "PLR0913", "PLR0915"]
 "backend/processing/pipeline_metrics.py" = ["PLR0913"]
 "backend/processing/speaker_merge.py" = ["C901", "PLR0912"]
